Quality of Life improvement! No changes to default config needed
The default config files no longer need to be modified.

The default config specifies paths to various datasets and other files. These have a "../" included which previously had to be removed as ReefModEngine.jl targeted a directory at the same level as its "root" .

Instead, a directory one level below is now targeted.